Rename edit screen to card_list
