+++ flutter/lib/views/card_list/card_list.dart
@@ -6,8 +6,8 @@ import 'package:delern_flutter/models/deck_access_model.dart';
 import 'package:delern_flutter/models/deck_model.dart';
 import 'package:delern_flutter/view_models/edit_deck_bloc.dart';
 import 'package:delern_flutter/views/base/screen_bloc_view.dart';
-import 'package:delern_flutter/views/edit_deck/deck_settings_widget.dart';
-import 'package:delern_flutter/views/edit_deck/scroll_to_beginning_list_widget.dart';
+import 'package:delern_flutter/views/card_list/deck_settings_widget.dart';
+import 'package:delern_flutter/views/card_list/scroll_to_beginning_list_widget.dart';
 import 'package:delern_flutter/views/helpers/arrow_to_fab_widget.dart';
 import 'package:delern_flutter/views/helpers/card_background_specifier.dart';
 import 'package:delern_flutter/views/helpers/edit_delete_dismissible_widget.dart';
